requirements.txt
flask[async]==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
pandas==2.1.0
openpyxl==3.1.2
requests==2.31.0
plotly==5.18.0
anthropic==0.43.0
python-json-logger==2.0.7
